Organizing imports in the quicksight module.
