+++ docs/source/what.rst
@@ -1,7 +1,7 @@
 What is AWS Data Wrangler?
 ==========================
 
-An `open-source <https://github.com/awslabs/aws-data-wrangler>`_ Python package that extends the power of `Pandas <https://github.com/pandas-dev/pandas>`_ library to AWS connecting **DataFrames** and AWS data related services (**Amazon Redshift**, **AWS Glue**, **Amazon Athena**, **Amazon EMR**, etc).
+An `open-source <https://github.com/awslabs/aws-data-wrangler>`_ Python package that extends the power of `Pandas <https://github.com/pandas-dev/pandas>`_ library to AWS connecting **DataFrames** and AWS data related services (**Amazon Redshift**, **AWS Glue**, **Amazon Athena**, **Amazon EMR**, **Amazon QuickSight**, etc).
 
 Built on top of other open-source projects like `Pandas <https://github.com/pandas-dev/pandas>`_, `Apache Arrow <https://github.com/apache/arrow>`_, `Boto3 <https://github.com/boto/boto3>`_, `s3fs <https://github.com/dask/s3fs>`_, `SQLAlchemy <https://github.com/sqlalchemy/sqlalchemy>`_, `Psycopg2 <https://github.com/psycopg/psycopg2>`_ and `PyMySQL <https://github.com/PyMySQL/PyMySQL>`_, it offers abstracted functions to execute usual ETL tasks like load/unload data from **Data Lakes**, **Data Warehouses** and **Databases**.
